pyproject.toml
[project]
name = "web-check"
version = "0.1.1"
description = "Docker-based security scanning toolkit"
authors = [
    {name = "Kevin De Benedetti", email = "contact@kevindb.dev"},
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.44.0",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "httpx>=0.28.1",
    "httpx-secure>=1.2.0",
    "structlog>=25.5.0",
    "python-json-logger>=4.1.0",
    "sqlalchemy>=2.0.49,<2.1",
    "alembic>=1.18.4",
    "aiosqlite>=0.20.0",
    "python-owasp-zap-v2.4>=0.0.22",
    "sslyze>=6.3.1",
    "sqlmap>=1.10.3,<1.10.4",
    "typer>=0.12.0",
    "rich>=15.0.0",
    "dnspython>=2.8.0",
    "kubernetes>=31.0.0",
    "python-dotenv>=1.2.2",
    "questionary>=2.1.1",
    "urllib3>=2.0",
]

[project.scripts]
web-check = "cli.main:app"
my-check = "my_check.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "ty>=0.0.31",
    "prek>=0.3.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
no-binary-package = ["sqlmap"]

[tool.hatch.build.targets.wheel]
packages = ["apps/api", "apps/cli", "apps/my_check"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by ruff format)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ty.environment]
python-version = "3.12"
python = ".venv"
python-platform = "all"
root = ["apps/api"]

[tool.ty.rules]
# Critical errors
possibly-unresolved-reference = "error"
unsupported-operator = "error"
division-by-zero = "error"

# Warnings (FastAPI/Starlette middleware typing issues)
invalid-argument-type = "warn"
unused-ignore-comment = "warn"
redundant-cast = "warn"

[tool.ty.src]
include = ["apps/api"]
exclude = ["**/__pycache__", "**/node_modules", ".venv", "outputs"]
respect-ignore-files = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["apps/api/tests"]
markers = [
    "slow: marks tests as slow (deselected by default in CI)",
    "integration: marks tests as integration tests",
]
addopts = "-v"
